[FIX] web: appswitcher layout
Apps which are not defining an icon to use as the appswitcher icon
were using a fa-cube element whose font-size was computed so that
the resulting fa-cube had the same size as the other app icons.

The problem is that when zooming on some browser, the font-size
computation is done wrong and as the fa-cube element has not the exact
same size as the other icons, the appswitcher layout is broken.

Solution is to get rid of the distinction and use a default image
instead of a fa-cube element. This also simplifies the LESS.