[FIX] web: don't open dialog before its content is loaded
Before this rev., when an action with target='new' was executed, its dialog was
directly opened, without waiting for the action to be loaded. This caused a
traceback in debug mode as the debug manager requires the view's id and type to
render, but only waits for the dialog to be open to do so.

A 'nice-to-have' side-effect of this commit is that the dialog now appears with
is content, instead of appearing empty, and then expanding when its content was
appended like before.