[FIX] l10n_us_check_printing: width not on body
The CSS of check reports were now added in a common report assets bundle
so we can't modify the parent `<body>` from a custom report as was done
in 1817da3.

This commit use an invisible block to set the 8.5in width (8.25 +
default left margin) and the 15px padding-left and padding-right to
`<body>` in report has to be countered.

This is crappy since we have to manage:

- body margins we can't touch,
- margin set in Odoo configuration,
- scaling

In master a solution where the body margins are removed should be done
instead but that can't be done in a stable version.

opw-669636

note: a better solution should be brought up in master.