[IMP] pos_blackbox_be: POSBox registration
Instead of registering databases we instead want to register
POSBoxes. This is easier to work with a technical point of view and it
also allows to have only a subset of POS configs use the Fiscal Data
Module.

Accomplishing this is somewhat tricky. In order to avoid an enormous
amount of if statements checking whether or not the current config is a
registered one or not we instead opted to include the blackbox assets
conditionally when loading the POS. This guarantees a familiar,
unmodified Point of Sale frontend when using a non-registered Point Of
Sale configuration.

On the backend we remove some view modifications and instead opt to
override the relevant functions to allow eg. editing of non-registered
orders but not of registered orders.