[FIX] website_sign: disable file input
When the user uploads a "large" PDF file, the RPC call to the method
`upload_from_dashboard` might take a while. Since it takes a few seconds
before the UI is blocked, it allows the user to quickly click again on
"Upload New PDF". This is not supposed to be allowed since one upload at a
time should be done. Moreover, this will generate a traceback when the
RPC is finished.

The fix disables the input file button before the RPC call, and enables it
upon failure or success.

opw-669235