[FIX] inter_company_rules: warehouse visibility & mandatory
Since the revision 9527178b186c9430cd02c31eb2e24e8d9ef15075,
the warehouse of the company is set on the sales orders as well
(Because, as mentioned, otherwise, it used the warehouse
from the other company).

The warehouse should therefore be visibile & required
in the configuration as soon as `so_from_po` or
'po_from_so` is checked.

opw-669795