[FIX] web: fix race condition in dataset
Dataset readslice alters ids after it is done.  However, rpc calls are
not,  ever, in any universe, guaranteed to return in the same order.
So, for example two consecutive search could arrive in the opposite
order, and the dataset will be leftover with the ids of the first query,
but the result of the second query might be used by the caller (case in
point, the kanban view)