[FIX] stock_barcode: initial mode
When the uses the `stock_picking_wave` module, opening a picking from a
wave gives a traceback which is caused by `stock_barcode`. Indeed, the
PickingBarcodeHandler widget is loaded in the picking form view thanks
to the field `_barcode_scanned`.

At the initialisation of the widget, it is assumed that the barcode
handler is included in a form view which is opened from an action. In
the case of the `stock_picking_wave` module, this is not the case.
Indeed, this is a dialog window appended to a form view.

Fixes https://github.com/odoo/odoo/issues/11123
opw-670649