[IMP] pos_blackbox_be: implement hash chain
This is a government request. For on-premise installations, the
government wants us to maintain a chain of hashes that should make
tampering with data directly in the database more difficult. The chain
goes chronologically through pro forma and non pro forma pos.orders. It
is computed as follows:

H_n = SHA1(H_(n-1) + PLU_n) & (2^64 - 1)

PLU_n refers to the 64 least significant bits of the PLU hash (cast to a
string) of the current order.