[IMP] website_crm_score: pass lead_id directly and not subselect to improve perf.
The explain analyze from psql show that it can influence the order.
So, if max_id - lead_id is small, it can apply the main request, only on these new lead (bigger than lead_id).
That will avoid to compute the main request on all the leads and after the computation remove all lead < lead_id.

To be efficient, it seems that it will be better to execute the scheduled action with small interval.
And so having the max_id - lead_id small.