[IMP] web: add option to split assets in debug manager
Now, the debug mode does not split assets by default.  This is great,
but sometimes we actually want to split assets.  This commit adds an
option in the developer toolbar to activate 'Assets Debugging', i.e.,
splitting and not minifying assets.

Also, a small improvement to the about menu (the option to activate
debugger menu only appears when not in debugging mode)