[IMP] web: synchronize control panel hiding with action displaying
When coming back to a previous action that didn't need a control panel using
the breadcrumbs, the control panel was hidden directly, before the action was
restored (which might be asynchronous). This produced some flickering.

With this rev., the control panel is hidden when the previous action is ready
to be restored.