[FIX] account_reports: bank reconciliation currencies
When a bank account is within a foreign currency,
the amounts in its bank statements are within the
bank account currency.

In other words: `account.bank.statement.line`.`amount`
is within the bank account currency.

In addition, the amounts displayed by the bank reconciliation
report (latest statements report) should be displayed
in the bank account currency, so the user can compare
easily these amounts with the amounts displayed
in the statements of his bank.

Therefore, when the bank account use a foreign currency,
the amounts must be displayed within this currency.
Therefore:

- `amount_currency` must be use instead of `balance`
   when the amounts are taken from the move line
   when the bank account use a foreign currency,
   as `balance` always uses the company currency,
   while `amount_currency` uses the journal currency.
- the currency symbol displayed must taken from the bank account
  journal currency when the bank account use a foreign currency,
  otherwise from the journal company currency.

opw-671534