[IMP] website_crm_score: add log to help debug + imp perf
- Help debug: keep track
We can know which lead has been merged with others in the log.

- Imp perf:
If we find less that BUNDLE_LEADS leads for a team, we dont re-check
the next time if new lead available for this team.
Else, if you have 1 lead for the same saleteam each 20 sec and that
the assignation to saleteam loop is bigger that 20 sec, you will redo
the job for one lead each time.

Previous commit introducted a small bug which make the assignation slow
We was rescoring all leads (lead_done) and not only the merged_leads.
So lead without merge was rescored for nothing.