[FIX] web: close appswitcher when doing action
A record can be followed through a channel, then with the channel
opened as IM the upper line of a message would be something like:

 Administrator - 24 minutes ago on [Clickable record name]

If we click on the link, we get to the record but that doesn't work when
on the app switcher.

This commit closes the appswitcher if it was opened when a window action
in current window (excluding popup) is being applied.

To do this, the action is returned from action manager's do_action.

closes #296
opw-667062