[FIX] web: regression do_action should return deferred
In 33ae4ff the action was set as a result of the done action deferred,
but some action currently don't return a deferred
(eg: ir_actions_report_xml defined in report module).

This commit add a $.when() deferred in this instance so a deferred will
always be available (either a directly resolved one or the one returned
by the action).

opw-673212
opw-673219