[FIX] website_contract: pay button re-render of recurring payments
This is related to revision
odoo/odoo@15f27b2614a4d6f86ded20fc7b3d8e796a2db078

When rendering the acquirer button
for a recurring payment,
the type of the transaction must be
`form_save` instead of simply
`form`.

Since the above revision, this was no longer
the case, because the transaction
type was hardcoded to `form` in the method
of the base controller route
`/quote/<int:order_id>/transaction/<int:acquirer_id>`
and thhis route was not overriden in `website_contract`
in order to change this value to `form_save`,
as it was done for the route
`/quote/<int:order_id>`

A new method `_get_payment_type` has been defined
in the `sale.order` itself, in the community module
`website_quote`, at the revision
odoo/odoo@895e77a336a4c6c6766f25cee7eb1a98869ba2af
to determine the transaction type from the `sale.order` directly,
in order to factorize the code and avoid copy/pasting
the button rendering code, while handling
the fact the transaction type should
be `form_save` for recurring payments.