[FIX] account_reports: available companies
With such a comparison, in a multicompany environment, it is likely that
the selected companies of the report are not strictly equal to the
companies available to the user. Therefore, a new report context is
created every time the user changes the search filters. The issue is
that this new context only includes the last filter specified by the
user (`given_context`), not all filters applied to the report.

opw-673762