[IMP] website_crm_score: extend scoring rule to allow to archive, unlink or score the domain
That allow to delete or archive lead before to score.
Until now, the solution was to score bad leads with a big negative score and manually delete them.
The same for spam e.g. we was scoring lead that will be deleted later manually by custom ir.filter.

Now we clean leads before to score automatically.