[FIX] web: FieldBinaryFile: update filename when switching records
The value of a FieldBinaryFile is the size of the uploaded file, not the
filename. But it had a 'filename' property, set at initialization or when
uploading a file.

However, this property wasn't updated when switching between records, so it
occured that the displayed filename wasn't the one of the visible record, but
the one of the first opened record.

A fix would be to update the 'filename' property in 'render_value()', as it is
called at each record switch. However, setting a widget's property shouldn't
be the concern of a 'render' function. We instead remove this 'filename'
property (as it was only used for rendering purpose) and directly use the
value of the corresponding 'filename' field.