[REF] web, web_enterprise: stop duplicating web
Removal of web addon and creation of web_enterprise addon, containing the
minimal stuff for the enterprise webclient to keep its look and be responsive.

The web_enterprise addons mainly contains:
  - JS code for the webclient, menu and appswitcher
  - Extensions of some community widgets for the views and control panel
    to be responsive
  - Less style of the enterprise webclient (mainly defined by variables and
    bootstrap override)
  - Less media queries for the webclient to be responsive
  - Some XML templates of mobile|desktop specific widgets

Debug manager has been uniformized between community and enterprise and it is
now available in the navbar's systray (and in modal headers for actions with
target='new').