[FIX] web: don't freeze on form warning escape
When the warning on form view is closed by cancel or OK button, all is
well. But if the modal is closed by any other way the def will stay pending
blocking next operations.

This commit set the deferred as dejected if not resolved when the modale
is closed.

closes odoo/odoo#11688
opw-671371

note: side port of odoo/odoo@6b76b08