[FIX] web: consistency between list and export
Following commit
https://github.com/odoo/odoo/commit/8276fb585371e78a500d57ff61f23ce10db24e9c,
the data exported should match the data
displayed on the list view. Therefore, we need to make sure that the
context used match the context of the list view.

When a filter is applied in the list view, the context is evaluated
thanks to pyeval (see the call to `/web/dataset/search_read`). The fix
is to evaluate the context in a similar way.

opw-674473