[FIX] Timesheet app: Don't redirect user after sync
A new event 'after_sync' can be triggered at the end of the synchronization process,
allowing to perform different actions based on the context at the time.
The user will no longer be redirected to the main screen nor will the menu be closed
at the end of the synchronization, as this could be confusing.