[IMP] sale_contract,website_contract: SOs subscription management field
In the Subscription apps, the Renewal and Upsell button
generate Sales Orders. In the case of the renewal, the
SO is renamed so that it looks like "SO2016/123456 -
Renewal", which make it easy to spot it in a list view.
For upselling there is no such mechanism yet.

Instead of appending suffixes to SOs (which pollutes the
view IMHO), we could add a selection field on SOs the
would contain the values "Upsell" and "Renewal"; this
field would be displayed in the list view of Sales Orders
when accessed through a Subscription (when cliking on the
Sales stat button) but hidden in other views. This would
also allow to remove a boolean field "update contract"
that is currently used to wipe contract when a renewal
SO is confirmed.