[FIX] web: remove FastClick from assets_backend
Fun fact: since FastClick was included in 9.0, it was actually not
loaded: the code in web client wait for a DOMContentLoaded event,
but the web client is instantiated only after DOMContentLoaded.

So, this commit simply remove FastClick from the assets. It cannot
simply be activated, because it might cause side effects.