[FIX] web_enterprise: prevent detaching twice webclient content
It could happen (e.g. by quickly clicking several times on the 'go back'
button of the browser) that the webclient content was detached from the DOM
while already being detached.

This is an issue because we store the detached nodes, so that we can re-attach
them later. So if we tried to detach an already detached content, we erased it
with some empty nodes, and it could thus never be re-attached (leaving the user
with the appswitcher only).