[ADD] web_enterprise: Menuitems search bar on the app switcher
A search bar is displayed under the navigation bar, where an fuzzy-search
is made each time a character is typed, hiding the app icons that are not
matching the seearch and displaying the secondary menuitems that are matching
the search. The user can use the navigation keys to travel along the menuitems
(Right, Left, Up, Down, Tab, Enter). Selecting a menuitem will make the user
jump to the related menuitem. Now it's much more easier to reach the 'Companies'
list view for example.

A shortcut has been added: Alt + h (sounds like 'home') that will display the
app switcher with the focus on the search bar to allow the user to type
directly the menuitem he's looking for. If the user press Alt + h again on the
app switcher, he'll jump back to the last view he was on.

NB: Following a fp request, the bar is not displayed until the first character
is typed.