[FIX] web_enterprise: Wrong app display
The $el variable was binded according to the menuitem action-id.

But, the application icon and the 1 secondary menuitem does have the
same action id (Sales app and Sales dashboard for instance).

When binding the variable $el, we thus make the distinction if the
menuitem is an app or a secondary menuitem.