[FIX] barcodes: fix tree view barcode handling
Oversight of 0f5cf67c4da40c9e39571c5757a33441b036dd3e. In a tree view
po_records will be a recordset with a find function. In the kanban
view it will be an Array of records.

Before 0f5cf67c4da40c9e39571c5757a33441b036dd3e this worked because both
an Array and the recordset provide a find function. After that commit
tree views will cause issues because it will call _.find on a recordset
object which won't work.

Although this is ugly, I'm not sure how else to do it as it is caused by
Kanban views and Tree views being inherently incompatible with each
other.