[FIX] web_enterprise: app_switcher: update url on action loaded
when clicking on an app icon.

Before this fix, the url was updated automatically because everything was
handled by the href. Unfortunately, if something went wrong when loading the
action (e.g. connection lost), the url was already modified so clicking on the
app again when the connection was back didn't work anymore (one needed to open
another action to make the former one available again).