[FIX] timesheet_grid: filter timesheet lines
account.analytic.line objects are actually created by and used for way
more than just timesheets. Which timesheet grid previously ignored
completely.

The way to know that an account.analytic.line is a timesheet line is
that it has a project_id set (according to @tac-odoo anyway).

Since that's rather unobvious, add an ``is_timesheet`` field (computed
and searchable but not stored) for code clarity and simplicity, then add
that to a bunch of search domains so the timesheet grid doesn't display
strange records coming from e.g. invoices.