[IMP] web_enterprise: garbage collecting of detached elements
When toggling the app switcher, some elements were left in the DOM on purpose,
because once detached, they can't be garbage collected (e.g. ui-autocomplete).

This rev. appends the detached elements in a document fragment, so that
garbage collecting correctly removes some elements from the fragment, and those
elements aren't attached back with the rest of the application.