[IMP] web: better renewal mechanism
Before redirecting to our website for the renewal,
trigger a ping and check if the db is not valid.

This case will happen when a subscription is closed
on our end, for example. To avoid having to dig too
much into the system, simply tell the user to click
on the "renew" link and that the local instance will
check everything with the ping.