[FIX] website_sign: IE11 TypeMismatchError
The FileReader.readAsDataUrl is throwing a TypeMismatchError for
no reason on IE11 (it does not do that on IE10 and on Edge).

This is a known problem in IE11, so as the error does not break
anything, this commit catches it and warns it in the console but
does not trigger a backend traceback which makes no sense to the user.