[IMP] project_forecast: reverse relation between time and resource_hours
"time" (the percentage of the period being "taken up") doesn't really
make for a good reference value, it makes more sense as an
optional (based on a resource being set up for the user) load indicator.

Thus make resource_hours (number of hours forecasted) the reference
value for a forecast, and compute the time (load percentage) if a
resource calendar has been set up for the user.