[FIX] web: prevent search dropdowns to close on click inside
In the Filters, Group By and Favorites dropdowns, when
clicking outside the `li` elements but still inside the
dropdown menu itself (like on the dividers, or next to the
advanced search `Apply` button), the dropdowns was closed.

This behavior was annoying when doing an advanced search
with a field using the datetime picker, as, to close the
datetime picker, you have to click outside it, and when you
did, it closed automatically the filter dropdown... So you
had to open it again to finally apply the filter.

opw-674310