[FIX] account_reports: comparison date consistency
Some reports (e.g. Balance Sheet) allow the comparison of data between
two periods. When a custom date is selected, and the comparison is set
to 'Previous Period', an inconsistency may arise between the comparison
date displayed in the report and the comparison date taken into account
server side. This is the case if the date chosen is July 31st and the
comparison is set to the previous period. The date displayed as a
comparison will be June 30th, while the date taken into account server
side is July 1st.

The fix is to make consistent the way the client and the server compute
this date.

opw-677657