[FIX] website_sign: slow search of recipient/responsible
When the database contains thousands of partners, the search for a
recipient or a responsible can become very slow.

This is because the `matcher` function is applied to all elements of the
partner list:
- when clicking on the field (the drop-down list is opened)
- when a character is entered in the field

The simplest solution is to activate the search if at least a given
number of characters (in this case, 3) is written in the field. Thanks
to this:
- no search is performed when clicking on the field
- the search is only performed when a meaningful amount of characters is
  written

Another solution would be to call the `name_get` on the `res.partner`
model (and potentially use the `delay` option of the ajax call).
However, this would require a major refactoring of the code.

opw-678222