[FIX] inter_company_rules: taxes with inter company user other than SUPERUSER
When configuring the inter company rules
to use another user than the administrator,
a user within the company which is the subject
of the inter company rule,
the taxes were missing in the invoice lines
because:
 - `_onchange_product_id` returns nothing
   if the line is not assigned to an invoice:
   ```
   if not self.invoice_id:
       return
   ```
 - and, since the `new` call was done with the current user,
   not as sudo:
   ```
   line2 = self.env['account.invoice.line'].new(inv_line_data)
   ```
   the assignation
   ```
   line2.invoice_id = invoice.id
   ```
   did not work, because it attempted to assign to the line
   an invoice to which the user of the line does not have access
   to

Therefore, in such a case, the `_onchange_product_id` returned
nothing, and no taxes were set on the lines of the invoice.

e.g:
 - Company A - User A (not superuser)
 - Company B - User B (not superuser)
 - Company B is set to create an invoice if the customer/supplier
   of an invoice of another company is Company B,
   as the User B

User A creates a customer invoice with Company B as customer, and validates it:
 - The invoice was correctly created as User B:
   `invoice = self.with_context(context).sudo(intercompany_uid).create(invoice_vals) `
 - But the call to `new` on `account.invoice.line` not:
   `line2 = self.env['account.invoice.line'].with_context(context).new(inv_line_data)`
   making it belonging to the User A, who does not have access to the `invoice`
   that was just created by the User B for the Company B
 - The assignation `line2.invoice_id = invoice.id` does not work
 - `_onchange_product_id` returns nothing, and the taxes are not set
   on the invoice which was created automatically.

The invoice lines must be instanciated with the same user
the invoice was created with, so the assignation can work,
and the taxes can be set.

opw-677691