[FIX] web_enterprise: correctly override /web controller
Unfortunately, controllers can't be overriden. There was an attempt to override
the '/web' controller in web_enterprise to remove the menu data from the
webclient rendering context, as the menus are rendered client side in
enterprise. This didn't work at all, and thus an unnecessary SQL request to
load the menu datas was performed each time someone accessed '/web'.