[FIX] project_forecast: Forecast by user with a lot of users
a `_group_by_full` is defined on the model forecast,
on its `user_id` field
meaning that if a read_group is applied on `user_id`,
it will return a result for all existing users,
even if they have no forecast.

This can obviously leads to performances issues if the
databases if full of users, especially customers account
(thanks to the ecommerce, forum, ...).

Besides, attempting to display the forecast of customers
is not really useful.

opw-681028