[FIX] web: pager minimum  on next/previous record
When browsing a form with a one2many / many2many field
with several pages, the one2many pager minimum was not reset
when going to the previous/next record, while the next/previous
record could not have multiple pages on its one2many/many2many,
and pager minimum which the user had on the former record could
be applied in the newly displayed record.

Besides, even if there is enough records in the list, there is no reason
to display this new record X2many value from its first page.

It's the case anyway when you edit/save, it comes back to the first page.

opw-679092