[FIX] website_version: multi website page cache key
The website pages (/page/*) have a mechanism of cache:
```
@http.route('/page/<page:page>', type='http', auth="public", website=True, cache=300)
```

In `ir_http.py` `dispatch` method, the page content are stored
in a dict with as key what is returned by the method `get_page_key`:

```
cache_time = getattr(func, 'routing', {}).get('cache')
cache_enable = cache_time and request.httprequest.method == "GET" and request.website.user_id.id == request.uid
cache_response = None
if cache_enable:
    key = self.get_page_key()
    try:
        r = self.pool.cache[key]
```

When working in multi website (different website pages
according to the domain) the website was not included
in this key, and therefore the cache could return the page content
of another website page if the page path was the same
e.g. 2 websites with a page `/page/test`

By adding the website id in the cache key, we
make sure the page content returned is the one
of the right website.

opw-677118