[FIX] web_enterprise: appswitcher padding-left
A big css padding-left alongside a align-items property changed to
flex-start was added to the appswitcher to prevent it to move when
a scrollbar is added or removed when an user type something in the
searchbar.

The problem was that chrome was not handling that change well in
the case some tour tips were attached to the appswitcher, making
them move randomly before reaching their final position.

This commit only change the padding-left by the size of the scrollbar
which is added or removed to fix the problem.
This also fix the problem of resizing the appswitcher.