[FIX] website_version: listing versions + create exp
Since qweb changes, the website_version module was not able to list
the existing versions since the data-view-xmlid attribute on the html
tag was replaced to be the name version of it instead of the numeric
version.

This commit changes the website_version code to get the numeric id
from the data-main-object attribute.