[FIX] account_reports: view invoice
On a report, when an account move line is linked to an invoice, clicking
on "View Invoice" always opens the invoice view of a supplier invoice.
However, it should open either the customer or the supplier view
depending on the invoice type.

Since the JS method `outboundLink` is able to retrieve the `view_id`,
and the Python method `get_model_id_and_name` already retrieves the
appropriate view, we should use this information to open the correct
view.

opw-682291