[FIX] website_contract: remove override of price in website sale cart
The initial purpose of this template was to see the discounted price in
the website sale cart and not the price unit.

But the commit odoo/odoo@045a8594144d1a759dae1f5d11e7814bb8f98434 fix it.

And removing this template, will remove a bug with website_event.
Where the price displayed in the cart is the product price and not the ticket price.