[FIX] account_bank_stmt_import_csv: fixes due to base_import migration
Migration brings error since method signature were not compatible anymore. Indeed,
`_parse_import_data` is api.multi now and `ids` parameters is required in old api.

After that, the override put some stuff in context to be used in calling method, which
is not compatible with new api. The usage of context variable is now moved from `do`
into `_parse_import_data`. The code is executed only when `statement_id` is passed
into context, which means the statement is already create, and can be modified, like
it was in `do`method.

To be cleaner, this fix does not modified context, it directly set the values to update
in a proper dict.