[FIX] website_enterprise: session.debug value
Now that the debug URL parameter can be equal to "assets", the session
debug variable had to be updated to not only be true or false. Now, its
value is false if the debug parameter is not specified in the url and
a the URL parameter string value otherwise (or 1 if the value is an
empty string).

This commit modifies the backend redirection which kept the debug mode
to also keep the =assets. Use the $.param.querystring function.