[FIX] account_reports: customers overdue, context_all
The variable `context_all` is only available
in the report `account.report.context.followup.all`,
not in the report 'single' report `account.followup.report`.
which is the report used when clicking on the button
"Overdue" on a partner form.

As a result, clicking on the button "Overdue"
leaded to a crash:
`'NoneType' object has no attribute 'partner_filter'`