[FIX] web_enterprise: remove useless breaking-tip rule
Commit 45579212ccf4fdef79408e5c268449ccfc7d0c36 removed the overflow auto rule on form view (which was
useless since it is the action manager (o_content) which scrolls) but
forgot to remove it from the no sheet form view.

On top of being useless, this rule now breaks the behavior when a tip
is placed on a field in an editable list, so this commit removes it.