[FIX] web: allow to recheck subscription status from alert panel and clarify wording
In some cases, it is possible for the local database to stay blocked
while the subscription on odoo.com has been updated and is valid again.
This is usually caused by the delay between subscription updates.
When the database is locked, it becomes quite difficult to force
an update.

To solve this, this commit simply adds a button on the panel that
allows forcing an update.

Also improve some button wording, for example "Add more users" was
understood by most users as "add users to your system", while
it meant "Add users to your subscription".

Closes #446