[FIX] project_forecast: Access error while trying to access project/tasks forecast
Clicking on menu Forecast in Project module, fetch the projects and tasks linked
to the forecast projects.

Due to two record rules: "Task Analysis multi-company" and "Project: multi-company",
it's not allowed for a user to see projects and tasks which aren't in his company.

Then a record rule for the project forecast is needed to filter the forecast project
linked to project or task not included in the company of the user.

opw:683976