[FIX] account_contract_dashboard: cohort with locale
When the user's language is not english, strptime returns an encoding error.

The fact that read_group also formats the date in this locale is a problem if any post-processing
is needed afterwards (with strptime here).
_read_group_raw returns (range, label) and not only label with range like date1/date2.
We can now use this date1 to compute the cohort date and post-proccess without any worry.

opw-685608