[FIX] app_switcher: quick search: traceback when no result and enter
In the `update` method called at each keypress, we try to find which app
is focused according to the focus index, even if the focus index is
null (meaning that nothing is selected). This results in a traceback
when you type enter when no results. We now only run the mechanism that
try to find the app is focused if something is focused.